+++ dart/dart.hpp
@@ -34,7 +34,6 @@
 #include "dart/config.hpp"
 #include "dart/common/common.hpp"
 #include "dart/math/math.hpp"
-#include "dart/integration/integration.hpp"
 #include "dart/collision/collision.hpp"
 #include "dart/lcpsolver/lcpsolver.hpp"
 #include "dart/constraint/constraint.hpp"