+++ dart/simulation/World.cpp
@@ -43,7 +43,7 @@
 #include "dart/constraint/BoxedLcpConstraintSolver.hpp"
 #include "dart/constraint/ConstrainedGroup.hpp"
 #include "dart/dynamics/Skeleton.hpp"
-#include "dart/integration/SemiImplicitEulerIntegrator.hpp"
+#include "dart/math/SemiImplicitEulerIntegrator.hpp"
 
 #include <iostream>
 #include <string>