+++ dart/simulation/World.hpp
@@ -57,9 +57,9 @@
 
 namespace dart {
 
-namespace integration {
+namespace math {
 class Integrator;
-} // namespace integration
+} // namespace math
 
 namespace dynamics {
 class Skeleton;